Refactor, bugfixed the formsets which can now give life to related objects
git-svn-id: svn://francepremium.nfrance.com/madeleinemarket.com@998 68bcc517-35a1-de11-95e8-003048d8a48a

+++ mad/model/model.php
@@ -437,13 +437,17 @@ class madModel {
     }
 
     public function delete( madBase $data, $cascade = true ) {
+        if ( !isset( $data['id'] ) ) {
+            throw new madModelExceptedId(  );
+        }
+        
         if ( $cascade ) {
             $sql = 'delete from mad_model where id = ' . self::ENCODE_ID_ENTITY . 
                 ' or attribute_value = :id';
         } else {
             $sql = 'delete from mad_model where id = ' . self::ENCODE_ID_ENTITY;
         }
-              
+        
         $statement = $this->query( $sql, array( 
             ':id' => $data['id'],
         ) );
